+++ CHANGES.rst
@@ -4,6 +4,10 @@ Changelog
 1.0 (unreleased)
 ----------------
 
+- Use index instead of template attribute in BatchView to be able to customize
+  only the template.
+  [vincentfretin]
+
 - Fixed wrong msgid for 'First item'.
   [vincentfretin]